Scripts folder structure is modified. Added new Character behaviour to inherit common facility. Added new Signal to boss die. method to get damage for the boss

+++ Assets/Scripts/DataStructure/Signals/BossAttackSignal.cs
@@ -1,3 +1,3 @@
 public class BossAttackSignal {
-	public uint damage;
+	public float damage;
 }
